Add get_binary_file_attachment_for template tag in difftags.py, and display basic FileAttachment information inline in the diff_file_fragment template.
This template tag will query for the FileAttachment based on the provided file,
which contains a filediff, and set the retrieved binary file attachment to a
variable whose name is provided as an argument to this tag. If no matching
FileAttachment is found or if there is more than one FileAttachment associated
with one FileDiff, None is returned, and an error is logged in the latter case.